Move and rename the helper class to the `ConstraintViolations.ExceptionBuilder`. Remove unused classes.
