+++ client/src/main/java/org/spine3/validate/ConstraintViolations.java
@@ -20,7 +20,16 @@
 
 package org.spine3.validate;
 
+import com.google.protobuf.Message;
+import com.google.protobuf.ProtocolMessageEnum;
+import com.google.protobuf.Value;
+import org.spine3.annotations.Internal;
+import org.spine3.base.Error;
+import org.spine3.base.ValidationError;
+import org.spine3.type.MessageClass;
+
 import java.util.List;
+import java.util.Map;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static java.lang.System.lineSeparator;
@@ -127,5 +136,90 @@ public class ConstraintViolations {
                          .append(childViolationFormatted);
         }
         return resultBuilder.toString();
+    }
+
+    /**
+     * A helper class for building exceptions used to report invalid {@code Message}s,
+     * which have fields that violate validation constraint(s).
+     *
+     * @param <E> type of {@code Exception} to build
+     * @param <M> type of the {@code Message}
+     * @param <C> type of the {@linkplain MessageClass MessageClass} of {@code |M|}.
+     * @param <R> type of an error code to use for error reporting; must be a Protobuf enum value
+     */
+    @Internal
+    public abstract static class ExceptionBuilder<E extends Exception,
+                                                  M extends Message,
+                                                  C extends MessageClass,
+                                                  R extends ProtocolMessageEnum> {
+
+        private final Iterable<ConstraintViolation> constraintViolations;
+        private final M message;
+
+        /**
+         * Creates an {@code ExceptionBuilder} instance for a given message and
+         * constraint violations,
+         *
+         * @param message              an invalid event message
+         * @param constraintViolations constraint violations for the event message
+         */
+        protected ExceptionBuilder(M message,
+                                   Iterable<ConstraintViolation> constraintViolations) {
+            this.constraintViolations = constraintViolations;
+            this.message = message;
+        }
+
+        /**
+         * Obtains a {@code MessageClass} for an invalid {@code Message}.
+         */
+        protected abstract C getMessageClass();
+
+        /**
+         * Obtains an error code to use for error reporting.
+         */
+        protected abstract R getErrorCode();
+
+        protected abstract String getErrorText();
+
+
+        protected abstract Map<String, Value> getMessageTypeAttribute(Message message);
+
+        protected abstract E createException(String exceptionMsg, M message, Error error);
+
+        private String formatExceptionMessage() {
+            return String.format("%s. Message class: %s. " +
+                                 "See Error.getValidationError() for details.",
+                          getErrorText(), getMessageClass());
+        }
+
+        private Error createError() {
+            final ValidationError validationError =
+                    ValidationError.newBuilder()
+                                   .addAllConstraintViolation(constraintViolations)
+                                   .build();
+            final R errorCode = getErrorCode();
+            final String typeName = errorCode.getDescriptorForType()
+                                             .getFullName();
+            final String errorTextTemplate = getErrorText();
+            final String errorText = String.format("%s %s",
+                                                   errorTextTemplate,
+                                                   toText(constraintViolations));
+
+            final Error.Builder error = Error.newBuilder()
+                                             .setType(typeName)
+                                             .setCode(errorCode.getNumber())
+                                             .setValidationError(validationError)
+                                             .setMessage(errorText)
+                                             .putAllAttributes(getMessageTypeAttribute(message));
+            return error.build();
+        }
+
+        /**
+         * Creates an exception instance for an invalid message, which has fields that
+         * violate validation constraint(s).
+         */
+        public E buildException() {
+            return createException(formatExceptionMessage(), message, createError());
+        }
     }
 }
